Move deprecated packages
