+++ deprecated/roosterjs-cross-window/lib/index.ts
