+++ deprecated/roosterjs-cross-window/package.json
