+++ deprecated/roosterjs-html-sanitizer/LICENSE
