+++ deprecated/roosterjs-plugin-image-resize/LICENSE
