+++ deprecated/roosterjs-plugin-image-resize/README.md
