+++ deprecated/roosterjs-plugin-image-resize/package.json
