+++ deprecated/roosterjs-plugin-picker/LICENSE
