+++ deprecated/roosterjs-plugin-picker/lib/index.ts
@@ -1,0 +1,4 @@
+// This package is deprecated.
+// All exports here are just for backward compatibility.
+// Please use type from their real packages directly
+export * from 'roosterjs-editor-plugins/lib/Picker';
