+++ deprecated/roosterjs-plugin-picker/package.json
