--- packages/roosterjs-plugin-image-resize/lib/index.ts
@@ -1,4 +1,0 @@
-// This package is deprecated.
-// All exports here are just for backward compatibility.
-// Please use type from their real packages directly
-export { ImageResize } from 'roosterjs-editor-plugins/lib/ImageResize';
